Changed price_rule_id to bigint

+++ db/migrate/20170508123128_add_price_rule_id_to_postcards.rb
@@ -1,5 +1,5 @@
 class AddPriceRuleIdToPostcards < ActiveRecord::Migration
   def change
-    add_column :postcards, :price_rule_id, :integer
+    add_column :postcards, :price_rule_id, :bigint
   end
 end
